build: add burger page and other dependencies
This commit will add two new directory namely recipes directory and img
directory. In the burger page or './recipes/burger.html' relative to the
index.hml, we add an image source that point to an image file in img
directory.

Also in the index.html, we update one of the links so that it point to
the burger page.